svg import : store font size as a double + some minor refactoring
git-svn-id: svn://scribus.net/branches/Version14x/Scribus@17435 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/plugins/import/svg/svgplugin.h
@@ -153,7 +153,7 @@ public:
 	QString FontStyle;
 	QString FontWeight;
 	QString FontStretch;
-	int FontSize;
+	double  FontSize;
 	QString FillCol;
 	QString fillRule;
 	QString GCol1;
